pyproject.toml
[project]
name = "sheriyans"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uuid>=1.30",
    "uvicorn>=0.40.0",
]
